pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "numpy>=2.4.3",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "scikit-image>=0.26.0",
]
